Bind to only localhost and use UInt32 (can send max 4GB at a time) as length.

+++ VOCR/Utils.swift
@@ -346,11 +346,11 @@ func predict() {
 	if let url = chooseFile() {
 		do {
 			let s = try Socket.create()
-			try s.connect(to: "localhost", port: 12345)
+			try s.connect(to:"localhost", port:12345)
 			let cicontext = CIContext()
 			let ciimage = CIImage(cgImage: loadImage(url)!)
 			let imageData = cicontext.jpegRepresentation(of: ciimage, colorSpace: ciimage.colorSpace!)
-			var length = imageData!.count
+			var length = UInt32(imageData!.count)
 			var data = Data(bytes: &length, count: MemoryLayout.size(ofValue: length))
 			data.append(imageData!)
 			try s.write(from: data)
